requirements.txt
django==4.1.5
djangorestframework==3.14
djangorestframework-camel-case==1.3.0
entrypoints==0.4
requests==2.28.1
gunicorn==20.1.0
django-cors-headers==3.13.0
pymemcache==4.0.0
django_redis==5.2.0
rq==1.11.1
celery==5.2.6
tzdata